Added custom queries route path
