+++ models/templates/player_points.html
@@ -3,7 +3,7 @@
 <meta charset="UTF-8">
 <title>Player stat</title>
 <body>
-    <form action="/form_player_stats" method="POST">
+    <form action="/form_player_points" method="POST">
         <p>Enter Player Name</p>
         <table>
             <tr>
@@ -14,7 +14,7 @@
                     <input type="text" name="player_name">
                 </td>
                 <td>
-                    <input type="submit" value="Here...">
+                    <input type="submit" value="Submit">
                 </td>
             </tr>
         </table>
